pyproject.toml
[project]
name = "keylol-tg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "climage>=0.2.2",
    "lxml>=5.4.0",
    "pyrofork>=2.3.67",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "tgcrypto>=1.2.5",
]
